resolve problems
